+++ src/atoms/iconButton/component.tsx
@@ -75,7 +75,8 @@ export const IconButton = (props: IIconButtonProps): React.ReactElement => {
   };
 
   const theme = useBuiltTheme('iconButtons', props.variant, props.theme);
-  const targetShouldOpenSameTab = props.targetShouldOpenSameTab || (props.targetShouldOpenSameTab == null && props.target && (props.target.startsWith('#') || props.target.startsWith('/')));
+  const isTargetWithinApp = props.target && (props.target.startsWith('#') || props.target.startsWith('/'));
+  const targetShouldOpenSameTab = props.targetShouldOpenSameTab || (props.targetShouldOpenSameTab == null && isTargetWithinApp);
   return (
     // @ts-ignore: as prop doesn't match type required
     <StyledIconButton
@@ -89,7 +90,7 @@ export const IconButton = (props: IIconButtonProps): React.ReactElement => {
       rel={props.target && 'noopener'}
       tabIndex={props.tabIndex || 0}
       target={props.target ? (targetShouldOpenSameTab ? '_self' : '_blank') : undefined}
-      as={props.target ? (isUsingCoreRouting && targetShouldOpenSameTab ? CoreLink : 'a') : undefined}
+      as={props.target ? (isUsingCoreRouting && targetShouldOpenSameTab && isTargetWithinApp ? CoreLink : 'a') : undefined}
     >
       {props.icon}
     </StyledIconButton>