+++ src/atoms/link/component.tsx
@@ -47,8 +47,10 @@ export interface ILinkProps extends IComponentProps<ILinkTheme> {
 
 export const Link = (props: ILinkProps): React.ReactElement => {
   const isUsingCoreRouting = useIsCoreRoutingEnabled();
+
   const theme = useBuiltTheme('links', props.variant, props.theme);
-  const shouldOpenSameTab = props.shouldOpenSameTab || (props.shouldOpenSameTab === undefined && props.target && (props.target.startsWith('#') || props.target.startsWith('/')));
+  const isTargetWithinApp = props.target.startsWith('#') || props.target.startsWith('/');
+  const targetShouldOpenSameTab = props.shouldOpenSameTab || (props.shouldOpenSameTab == null && props.target && isTargetWithinApp);
   return (
   // @ts-ignore: as prop doesn't match type required
     <StyledLink
@@ -58,8 +60,8 @@ export const Link = (props: ILinkProps): React.ReactElement => {
       href={props.isEnabled ? props.target : undefined}
       rel={'noopener'}
       tabIndex={props.tabIndex || 0}
-      target={shouldOpenSameTab ? '_self' : '_blank'}
-      as={isUsingCoreRouting && shouldOpenSameTab ? CoreLink : undefined}
+      target={targetShouldOpenSameTab ? '_self' : '_blank'}
+      as={isUsingCoreRouting && targetShouldOpenSameTab && isTargetWithinApp ? CoreLink : undefined}
     >
       {props.text}
     </StyledLink>