+++ src/atoms/linkBase/component.tsx
@@ -84,7 +84,8 @@ export const LinkBase = (props: ILinkBaseProps): React.ReactElement => {
   };
 
   const theme = useBuiltTheme('linkBases', props.variant, props.theme);
-  const targetShouldOpenSameTab = props.targetShouldOpenSameTab || (props.targetShouldOpenSameTab == null && props.target && (props.target.startsWith('#') || props.target.startsWith('/')));
+  const isTargetWithinApp = props.target && (props.target.startsWith('#') || props.target.startsWith('/'));
+  const targetShouldOpenSameTab = props.targetShouldOpenSameTab || (props.targetShouldOpenSameTab == null && isTargetWithinApp);
   return (
     // @ts-ignore: as prop doesn't match type required
     <StyledLinkBase
@@ -97,7 +98,7 @@ export const LinkBase = (props: ILinkBaseProps): React.ReactElement => {
       rel={props.target ? 'noopener' : undefined}
       tabIndex={props.tabIndex || 0}
       target={props.target ? (targetShouldOpenSameTab ? '_self' : '_blank') : undefined}
-      as={props.target ? (isUsingCoreRouting && targetShouldOpenSameTab ? CoreLink : 'a') : undefined}
+      as={props.target ? (isUsingCoreRouting && targetShouldOpenSameTab && isTargetWithinApp ? CoreLink : 'a') : undefined}
     >
       {props.children}
     </StyledLinkBase>