+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Fixed colors theme creation to calculate text and textOnBrand correctly when background is dark
 - [MINOR] Fixed `InputWrapper` to set placeholder on correct elements
 - [MINOR] Updated to react-markdown v6
+- [MINOR] Fixed `Button`, `IconButton`, `Link` and `LinkBase` to only use `CoreLink` for internal links
 
 ### Removed